fix: lets move the git credentials loading code into the file handler
